requirements.txt
# Core dependencies
httpx>=0.24.1        # HTTP client for downloading from archive.org
tqdm>=4.65.0         # Progress bars for downloads and batch processing
pydantic>=2.0.0      # Data validation and settings management
loguru>=0.7.0        # Improved logging
PyYAML>=6.0          # YAML configuration file support

# Local LLM
ollama>=0.1.0        # Python client for Ollama

# Testing
pytest>=7.0.0        # Test framework
pytest-cov>=4.1.0    # Test coverage reporting
pytest-mock>=3.10.0  # Mocking for tests

# Development tools
black>=23.3.0        # Code formatting
isort>=5.12.0        # Import sorting
mypy>=1.3.0          # Static type checking
pylint>=2.17.0       # Code linting

# Utilities
python-dateutil>=2.8.2  # Date parsing and manipulation
tenacity>=8.2.2         # Retry logic
psutil>=5.9.0           # Process and system utilities for benchmarking
